+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/CallbackQueryId.kt
@@ -7,7 +7,11 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class CallbackQueryId(
     val string: String
-)
+) {
+    override fun toString(): String {
+        return string
+    }
+}
 
 @Deprecated("Renamed", ReplaceWith("CallbackQueryId", "dev.inmo.tgbotapi.types.CallbackQueryId"))
 typealias CallbackQueryIdentifier = CallbackQueryId