+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineMessageId.kt
@@ -7,6 +7,10 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class InlineMessageId(
     val string: String
-)
+) {
+    override fun toString(): String {
+        return string
+    }
+}
 @Deprecated("Renamed", ReplaceWith("InlineMessageId", "dev.inmo.tgbotapi.types.InlineMessageId"))
 typealias InlineMessageIdentifier = InlineMessageId