+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/RawChatId.kt
@@ -1,0 +1,20 @@
+package dev.inmo.tgbotapi.types
+
+import kotlinx.serialization.Serializable
+import kotlin.jvm.JvmInline
+
+@Serializable
+@JvmInline
+value class RawChatId(
+    val long: Long
+) {
+    companion object {
+        val DefaultUserId = RawChatId(136817688L) // I do not know why, it is Telegram crutch
+    }
+
+    override fun toString(): String {
+        return long.toString()
+    }
+}
+@Deprecated("Renamed", ReplaceWith("RawChatId", "dev.inmo.tgbotapi.types.RawChatId"))
+typealias Identifier = RawChatId