+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ShippingQueryId.kt
@@ -7,7 +7,11 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class ShippingQueryId(
     val string: String
-)
+) {
+    override fun toString(): String {
+        return string
+    }
+}
 
 @Deprecated("Renamed", ReplaceWith("ShippingQueryId", "dev.inmo.tgbotapi.types.ShippingQueryId"))
 typealias ShippingQueryIdentifier = ShippingQueryId