+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/WebAppQueryId.kt
@@ -7,4 +7,8 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class WebAppQueryId(
     val string: String
-)
+) {
+    override fun toString(): String {
+        return string
+    }
+}